# specs/001-personal-tech-blog/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Tech Blog API
  description: |
    REST API for a personal tech blog platform with public content access
    and protected admin endpoints for content management.
  version: 1.0.0
  contact:
    name: Blog Owner
    email: admin@example.com

servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://yourblog.vercel.app/api
    description: Production

tags:
  - name: Public
    description: Publicly accessible endpoints (no authentication required)
  - name: Admin
    description: Protected admin endpoints (authentication required)
  - name: Auth
    description: Authentication endpoints

paths:
  # ===== PUBLIC ENDPOINTS =====

  /posts:
    get:
      tags: [Public]
      summary: Get all published blog posts
      description: Returns a list of all published blog posts, sorted by publication date (newest first)
      operationId: getPublishedPosts
      parameters:
        - name: limit
          in: query
          description: Maximum number of posts to return
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of posts to skip (for pagination)
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Successful response with list of published posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/BlogPostSummary'
                  total:
                    type: integer
                    description: Total number of published posts
                  limit:
                    type: integer
                  offset:
                    type: integer
        '500':
          $ref: '#/components/responses/InternalServerError'

  /posts/{slug}:
    get:
      tags: [Public]
      summary: Get blog post by slug
      description: Returns full details of a single published blog post
      operationId: getPostBySlug
      parameters:
        - name: slug
          in: path
          required: true
          description: URL-friendly post identifier
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
      responses:
        '200':
          description: Successful response with post details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPostFull'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /resume:
    get:
      tags: [Public]
      summary: Get all resume items
      description: Returns all resume items sorted by display order and date
      operationId: getResumeItems
      responses:
        '200':
          description: Successful response with resume items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResumeItem'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ===== AUTHENTICATION ENDPOINTS =====

  /auth/login:
    post:
      tags: [Auth]
      summary: Admin login
      description: Authenticate admin user and create session
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: admin@example.com
                password:
                  type: string
                  format: password
                  minLength: 12
                  example: SecurePassword123!
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: Session cookie (httpOnly, secure, sameSite=strict)
              schema:
                type: string
                example: sessionId=abc123; HttpOnly; Secure; SameSite=Strict; Max-Age=1800
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      email:
                        type: string
                      name:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Too many login attempts. Please try again in 15 minutes.
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/logout:
    post:
      tags: [Auth]
      summary: Admin logout
      description: Destroy admin session
      operationId: logout
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/session:
    get:
      tags: [Auth]
      summary: Get current session
      description: Check if admin is authenticated and get session info
      operationId: getSession
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Authenticated session
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
                    example: true
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      email:
                        type: string
                      name:
                        type: string
                  expiresAt:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
                    example: false

  # ===== ADMIN ENDPOINTS - BLOG POSTS =====

  /admin/posts:
    get:
      tags: [Admin]
      summary: Get all blog posts (including drafts)
      description: Admin-only endpoint to retrieve all blog posts
      operationId: getAllPostsAdmin
      security:
        - cookieAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [draft, published, all]
            default: all
      responses:
        '200':
          description: Successful response with all posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/BlogPostFull'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Admin]
      summary: Create new blog post
      description: Create a new blog post (draft or published)
      operationId: createPost
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlogPostInput'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPostFull'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: A post with this slug already exists
        '500':
          $ref: '#/components/responses/InternalServerError'

  /admin/posts/{id}:
    get:
      tags: [Admin]
      summary: Get blog post by ID (admin)
      description: Retrieve any blog post by ID (including drafts)
      operationId: getPostByIdAdmin
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPostFull'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Admin]
      summary: Update blog post
      description: Update an existing blog post
      operationId: updatePost
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlogPostInput'
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPostFull'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Admin]
      summary: Delete blog post
      description: Permanently delete a blog post
      operationId: deletePost
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Post deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ===== ADMIN ENDPOINTS - RESUME ITEMS =====

  /admin/resume:
    get:
      tags: [Admin]
      summary: Get all resume items (admin)
      description: Admin endpoint to retrieve all resume items
      operationId: getResumeItemsAdmin
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResumeItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Admin]
      summary: Create resume item
      description: Create a new resume item
      operationId: createResumeItem
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeItemInput'
      responses:
        '201':
          description: Resume item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /admin/resume/{id}:
    get:
      tags: [Admin]
      summary: Get resume item by ID
      description: Retrieve a single resume item
      operationId: getResumeItemById
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Admin]
      summary: Update resume item
      description: Update an existing resume item
      operationId: updateResumeItem
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeItemInput'
      responses:
        '200':
          description: Resume item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Admin]
      summary: Delete resume item
      description: Permanently delete a resume item
      operationId: deleteResumeItem
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Resume item deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

# ===== COMPONENTS =====

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionId
      description: Session-based authentication using httpOnly cookie

  schemas:
    BlogPostSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 255
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
        excerpt:
          type: string
          maxLength: 500
          nullable: true
        publishedAt:
          type: string
          format: date-time
        metadata:
          type: object
          nullable: true
          properties:
            tags:
              type: array
              items:
                type: string
            category:
              type: string
            readingTime:
              type: integer
      required:
        - id
        - title
        - slug
        - publishedAt

    BlogPostFull:
      allOf:
        - $ref: '#/components/schemas/BlogPostSummary'
        - type: object
          properties:
            content:
              type: string
              description: Full blog post content (Markdown or HTML)
            status:
              type: string
              enum: [draft, published]
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
          required:
            - content
            - status
            - createdAt
            - updatedAt

    BlogPostInput:
      type: object
      required:
        - title
        - slug
        - content
        - status
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
          minLength: 1
          maxLength: 255
        content:
          type: string
          minLength: 1
        excerpt:
          type: string
          maxLength: 500
          nullable: true
        status:
          type: string
          enum: [draft, published]
          default: draft
        metadata:
          type: object
          nullable: true
          properties:
            tags:
              type: array
              items:
                type: string
            category:
              type: string
            seo:
              type: object
              properties:
                metaDescription:
                  type: string
                ogImage:
                  type: string
            readingTime:
              type: integer

    ResumeItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        company:
          type: string
          minLength: 1
          maxLength: 255
        position:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          minLength: 1
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
          description: Null indicates current/ongoing position
        displayOrder:
          type: integer
          description: Higher values displayed first
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - company
        - position
        - description
        - startDate
        - displayOrder
        - createdAt
        - updatedAt

    ResumeItemInput:
      type: object
      required:
        - company
        - position
        - description
        - startDate
      properties:
        company:
          type: string
          minLength: 1
          maxLength: 255
        position:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          minLength: 1
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        displayOrder:
          type: integer
          default: 0

    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
      required:
        - error

  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation failed. Title is required.

    Unauthorized:
      description: Authentication required or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Authentication required. Please log in.

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Blog post not found

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: An unexpected error occurred